FixCommit2

+++ src/main/java/com/project/messenger/service/ChatService.java
@@ -88,6 +88,17 @@ public class ChatService {
     }
 
     @Transactional(readOnly = true)
+    public Chat getChatWithMembersUsersData(Long chatId, String email) {
+        User user = userService.findByEmail(email);
+        Chat chat = chatRepository.findChatWithMembersWithUsersById(chatId)
+                .orElseThrow(() -> new IllegalArgumentException("Чат не найден"));
+        if (!chatMemberRepository.existsByChatIdAndUserId(chatId, user.getId())) {
+            throw new SecurityException("У вас нет доступа к этому чату");
+        }
+        return chat;
+    }
+
+    @Transactional(readOnly = true)
     public User getChatContact(Long chatId, String currentUserEmail) {
         Chat chat = getChat(chatId, currentUserEmail);
         return chat.getMembers().stream()
@@ -162,6 +173,11 @@ public class ChatService {
     public void removeMemberFromGroup(Long chatId, Long userId) {
         ChatMember member = chatMemberRepository.findByChatIdAndUserId(chatId, userId)
                 .orElseThrow(() -> new IllegalArgumentException("Участник не найден в чате"));
+
+        if (member.isAdmin()) {
+            throw new IllegalStateException("Нельзя удалить администратора чата");
+        }
+
         chatMemberRepository.delete(member);
     }
